requirements.txt
# API Framework
fastapi>=0.104.1
uvicorn>=0.24.0

# Database
sqlalchemy>=2.0.23
alembic>=1.12.1
psycopg2-binary>=2.9.9

# LLM Integration
openai>=1.3.7
anthropic>=0.5.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.2
tenacity>=8.2.3
